+++ R/zzz.R
@@ -7,6 +7,7 @@
     ":=",
     "ADOSEDUR",
     "AFRLT",
+    "ANL01FL",
     "ARRLT",
     "ATPT",
     "ATPTN",